requirements.txt
streamlit==1.30.0
transformers==4.38.0
torch==2.2.0
numpy==1.26.3
matplotlib==3.8.2
seaborn==0.13.2
pandas==2.2.0
plotly==5.18.0
networkx==3.2.1
wordcloud==1.9.3
shap==0.44.0
pillow==10.2.0
